refactor: 체스 관련 dao를 하나로 관리하도록 수정
